requirements.txt
Flask==3.1.0
pyserial==3.5
gunicorn==23.0.0
waitress==3.0.0
